♻️ [Refactor] : Location 모듈 css 공통 css 로 통합
